+++ src/main/java/br/com/collei/lavi/morphology/enums/EnumHarmony.java
@@ -1,18 +1,21 @@
-package br.com.collei.lavi.morphology;
+package br.com.collei.lavi.morphology.enums;
 
 import java.util.Arrays;
 import java.util.List;
 
-import br.com.collei.lavi.morphology.filter.VerbConjugationFilter;
-
 /**
- * <p>Enumerated verb voices.</p>
+ * <p>Enumerated harmony types.</p>
  * @author almir jr.
  *
  */
-public enum EnumVerbVoices implements VerbConjugationFilter {
+public enum EnumHarmony {
 
-	ACTIVE(1,"ACTIVE"), MEDIAL(2,"MEDIAL"), PASSIVE(3,"PASSIVE");
+	BACK(1,"BACK"), FRONT(2,"FRONT");
+	
+	/**
+	 * <p>Fix 'bad class file' issues</p>
+	 */
+	private static final long serialVersionUID = 98274998904169L;
 	
 	/**
 	 * <p>numeric value of the type.</p>
@@ -29,7 +32,7 @@ public enum EnumVerbVoices implements VerbConjugationFilter {
 	 * @param value the numeric value of the type
 	 * @param description the textual description of the value
 	 */
-	EnumVerbVoices(int value, String description) {
+	EnumHarmony(int value, String description) {
 		this.value = value;
 		this.description = description;
 	}
@@ -45,6 +48,7 @@ public enum EnumVerbVoices implements VerbConjugationFilter {
 	/**
 	 * <p>gets the textual description of the type.</p>
 	 */
+	@Override
 	public String toString() {
 		return this.description;
 	}
@@ -53,8 +57,8 @@ public enum EnumVerbVoices implements VerbConjugationFilter {
 	 * <p>Generates the list of the enumerated types.</p>
 	 * @return the list of all possible values
 	 */
-	public static List<EnumVerbVoices> asList() {
-		return Arrays.asList(ACTIVE, MEDIAL, PASSIVE);
+	public static List<EnumHarmony> asList() {
+		return Arrays.asList(BACK, FRONT);
 	}
 	
 }
