+++ src/main/java/br/com/collei/lavi/morphology/enums/EnumNounCases.java
@@ -1,16 +1,32 @@
-package br.com.collei.lavi.morphology;
+package br.com.collei.lavi.morphology.enums;
 
 import java.util.Arrays;
 import java.util.List;
 
+import br.com.collei.lavi.morphology.enums.filter.NounDeclensionFilter;
+
 /**
- * <p>Enumerated harmony types.</p>
+ * <p>Enumerated noun cases.</p>
  * @author almir jr.
  *
  */
-public enum EnumHarmony {
+public enum EnumNounCases implements NounDeclensionFilter {
 
-	BACK(1,"BACK"), FRONT(2,"FRONT");
+	NOMINATIVE(1,"NOMINATIVE"),
+	ACCUSATIVE(2,"ACCUSATIVE"),
+	GENITIVE(3,"GENITIVE"),
+	DATIVE(4,"DATIVE"),
+	ABLATIVE(5,"ABLATIVE"),
+	LOCATIVE(6,"LOCATIVE"),
+	INSTRUMENTAL(7,"INSTRUMENTAL"),
+	PARTITIVE(8,"PARTITIVE"),
+	ABESSIVE(9,"ABESSIVE"),
+	COMITATIVE(10,"COMITATIVE");
+	
+	/**
+	 * <p>Fix 'bad class file' issues</p>
+	 */
+	private static final long serialVersionUID = 98274998904169L;
 	
 	/**
 	 * <p>numeric value of the type.</p>
@@ -27,7 +43,7 @@ public enum EnumHarmony {
 	 * @param value the numeric value of the type
 	 * @param description the textual description of the value
 	 */
-	EnumHarmony(int value, String description) {
+	EnumNounCases(int value, String description) {
 		this.value = value;
 		this.description = description;
 	}
@@ -39,20 +55,21 @@ public enum EnumHarmony {
 	public int getValue() {
 		return this.value;
 	}
-	
+
 	/**
 	 * <p>gets the textual description of the type.</p>
 	 */
+	@Override
 	public String toString() {
 		return this.description;
-	}
+	}	
 	
 	/**
 	 * <p>Generates the list of the enumerated types.</p>
 	 * @return the list of all possible values
 	 */
-	public static List<EnumHarmony> asList() {
-		return Arrays.asList(BACK, FRONT);
+	public static List<EnumNounCases> asList() {
+		return Arrays.asList(NOMINATIVE, ACCUSATIVE, GENITIVE, DATIVE, ABLATIVE, LOCATIVE, INSTRUMENTAL, PARTITIVE, ABESSIVE, COMITATIVE);
 	}
-	
+
 }
