+++ src/main/java/br/com/collei/lavi/morphology/enums/EnumNounPersons.java
@@ -1,18 +1,23 @@
-package br.com.collei.lavi.morphology;
+package br.com.collei.lavi.morphology.enums;
 
 import java.util.Arrays;
 import java.util.List;
 
-import br.com.collei.lavi.morphology.filter.VerbConjugationFilter;
+import br.com.collei.lavi.morphology.enums.filter.NounDeclensionFilter;
 
 /**
- * <p>Enumerated verb tenses.</p>
+ * <p>Enumerated possessive persons of nominative.</p>
  * @author almir jr.
  *
  */
-public enum EnumVerbTenses implements VerbConjugationFilter {
+public enum EnumNounPersons implements NounDeclensionFilter {
 
-	IMPERFECT(1,"IMPERFECT"), PERFECT(2,"PERFECT");
+	NEUTRAL(0,"NEUTRAL"), I(1,"I"), YOU(2,"YOU"), HE_SHE_IT(3,"HE_SHE_IT"), WE(4,"WE"), YOU_PLURAL(5,"YOU_PLURAL"), THEY(6,"THEY");
+	
+	/**
+	 * <p>Fix 'bad class file' issues</p>
+	 */
+	private static final long serialVersionUID = 98274998904169L;
 	
 	/**
 	 * <p>numeric value of the type.</p>
@@ -29,7 +34,7 @@ public enum EnumVerbTenses implements VerbConjugationFilter {
 	 * @param value the numeric value of the type
 	 * @param description the textual description of the value
 	 */
-	EnumVerbTenses(int value, String description) {
+	EnumNounPersons(int value, String description) {
 		this.value = value;
 		this.description = description;
 	}
@@ -45,6 +50,7 @@ public enum EnumVerbTenses implements VerbConjugationFilter {
 	/**
 	 * <p>gets the textual description of the type.</p>
 	 */
+	@Override
 	public String toString() {
 		return this.description;
 	}
@@ -53,8 +59,8 @@ public enum EnumVerbTenses implements VerbConjugationFilter {
 	 * <p>Generates the list of the enumerated types.</p>
 	 * @return the list of all possible values
 	 */
-	public static List<EnumVerbTenses> asList() {
-		return Arrays.asList(IMPERFECT, PERFECT);
+	public static List<EnumNounPersons> asList() {
+		return Arrays.asList(I, YOU, HE_SHE_IT, WE, YOU_PLURAL, THEY);
 	}
 	
 }
