+++ src/main/java/br/com/collei/lavi/morphology/enums/EnumNounPluralities.java
@@ -1,28 +1,24 @@
-package br.com.collei.lavi.morphology;
+package br.com.collei.lavi.morphology.enums;
 
 import java.util.Arrays;
 import java.util.List;
 
-import br.com.collei.lavi.morphology.filter.NounDeclensionFilter;
+import br.com.collei.lavi.morphology.enums.filter.NounDeclensionFilter;
 
 /**
- * <p>Enumerated noun cases.</p>
+ * <p>Enumerated noun pluralities.</p>
  * @author almir jr.
  *
  */
-public enum EnumNounCases implements NounDeclensionFilter {
+public enum EnumNounPluralities implements NounDeclensionFilter {
 
-	NOMINATIVE(1,"NOMINATIVE"),
-	ACCUSATIVE(2,"ACCUSATIVE"),
-	GENITIVE(3,"GENITIVE"),
-	DATIVE(4,"DATIVE"),
-	ABLATIVE(5,"ABLATIVE"),
-	LOCATIVE(6,"LOCATIVE"),
-	INSTRUMENTAL(7,"INSTRUMENTAL"),
-	PARTITIVE(8,"PARTITIVE"),
-	ABESSIVE(9,"ABESSIVE"),
-	COMITATIVE(10,"COMITATIVE");
+	SINGULAR(1,"SINGULAR"), DUAL(2,"DUAL"), PLURAL(3,"PLURAL");
 	
+	/**
+	 * <p>Fix 'bad class file' issues</p>
+	 */
+	private static final long serialVersionUID = 98274998904169L;
+
 	/**
 	 * <p>numeric value of the type.</p>
 	 */
@@ -38,7 +34,7 @@ public enum EnumNounCases implements NounDeclensionFilter {
 	 * @param value the numeric value of the type
 	 * @param description the textual description of the value
 	 */
-	EnumNounCases(int value, String description) {
+	EnumNounPluralities(int value, String description) {
 		this.value = value;
 		this.description = description;
 	}
@@ -50,10 +46,11 @@ public enum EnumNounCases implements NounDeclensionFilter {
 	public int getValue() {
 		return this.value;
 	}
-
+	
 	/**
 	 * <p>gets the textual description of the type.</p>
 	 */
+	@Override
 	public String toString() {
 		return this.description;
 	}	
@@ -62,8 +59,8 @@ public enum EnumNounCases implements NounDeclensionFilter {
 	 * <p>Generates the list of the enumerated types.</p>
 	 * @return the list of all possible values
 	 */
-	public static List<EnumNounCases> asList() {
-		return Arrays.asList(NOMINATIVE, ACCUSATIVE, GENITIVE, DATIVE, ABLATIVE, LOCATIVE, INSTRUMENTAL, PARTITIVE, ABESSIVE, COMITATIVE);
+	public static List<EnumNounPluralities> asList() {
+		return Arrays.asList(SINGULAR, DUAL, PLURAL);
 	}
-
+	
 }
