+++ src/main/java/br/com/collei/lavi/morphology/enums/EnumVerbModes.java
@@ -1,18 +1,23 @@
-package br.com.collei.lavi.morphology;
+package br.com.collei.lavi.morphology.enums;
 
 import java.util.Arrays;
 import java.util.List;
 
-import br.com.collei.lavi.morphology.filter.VerbConjugationFilter;
+import br.com.collei.lavi.morphology.enums.filter.VerbConjugationFilter;
 
 /**
- * <p>Enumerated verb persons.</p>
+ * <p>Enumerated verb modes.</p>
  * @author almir jr.
  *
  */
-public enum EnumVerbPersons implements VerbConjugationFilter {
+public enum EnumVerbModes implements VerbConjugationFilter {
 
-	I(1,"I"), YOU(2,"YOU"), HE_SHE_IT(3,"HE_SHE_IT"), WE(4,"WE"), YOU_PLURAL(5,"YOU_PLURAL"), THEY(6,"THEY");
+	FACTUAL(1,"FACTUAL"), DESIDERATIVE(2,"DESIDERATIVE"), IMPERATIVE(3,"IMPERATIVE");
+	
+	/**
+	 * <p>Fix 'bad class file' issues</p>
+	 */
+	private static final long serialVersionUID = 98274998904169L;
 	
 	/**
 	 * <p>numeric value of the type.</p>
@@ -29,7 +34,7 @@ public enum EnumVerbPersons implements VerbConjugationFilter {
 	 * @param value the numeric value of the type
 	 * @param description the textual description of the value
 	 */
-	EnumVerbPersons(int value, String description) {
+	EnumVerbModes(int value, String description) {
 		this.value = value;
 		this.description = description;
 	}
@@ -45,16 +50,17 @@ public enum EnumVerbPersons implements VerbConjugationFilter {
 	/**
 	 * <p>gets the textual description of the type.</p>
 	 */
+	@Override
 	public String toString() {
 		return this.description;
-	}
+	}	
 	
 	/**
 	 * <p>Generates the list of the enumerated types.</p>
 	 * @return the list of all possible values
 	 */
-	public static List<EnumVerbPersons> asList() {
-		return Arrays.asList(I, YOU, HE_SHE_IT, WE, YOU_PLURAL, THEY);
+	public static List<EnumVerbModes> asList() {
+		return Arrays.asList(FACTUAL, DESIDERATIVE, IMPERATIVE);
 	}
 	
 }
