+++ src/main/java/br/com/collei/lavi/morphology/enums/EnumVerbTenses.java
@@ -1,18 +1,23 @@
-package br.com.collei.lavi.morphology;
+package br.com.collei.lavi.morphology.enums;
 
 import java.util.Arrays;
 import java.util.List;
 
-import br.com.collei.lavi.morphology.filter.VerbConjugationFilter;
+import br.com.collei.lavi.morphology.enums.filter.VerbConjugationFilter;
 
 /**
- * <p>Enumerated verb definitenesses.</p>
+ * <p>Enumerated verb tenses.</p>
  * @author almir jr.
  *
  */
-public enum EnumVerbDefiniteness implements VerbConjugationFilter {
+public enum EnumVerbTenses implements VerbConjugationFilter {
 
-	INDEFINITE(1,"INDEFINITE"), DEFINITE(2,"DEFINITE");
+	IMPERFECT(1,"IMPERFECT"), PERFECT(2,"PERFECT");
+	
+	/**
+	 * <p>Fix 'bad class file' issues</p>
+	 */
+	private static final long serialVersionUID = 98274998904169L;
 	
 	/**
 	 * <p>numeric value of the type.</p>
@@ -29,7 +34,7 @@ public enum EnumVerbDefiniteness implements VerbConjugationFilter {
 	 * @param value the numeric value of the type
 	 * @param description the textual description of the value
 	 */
-	EnumVerbDefiniteness(int value, String description) {
+	EnumVerbTenses(int value, String description) {
 		this.value = value;
 		this.description = description;
 	}
@@ -45,6 +50,7 @@ public enum EnumVerbDefiniteness implements VerbConjugationFilter {
 	/**
 	 * <p>gets the textual description of the type.</p>
 	 */
+	@Override
 	public String toString() {
 		return this.description;
 	}
@@ -53,8 +59,8 @@ public enum EnumVerbDefiniteness implements VerbConjugationFilter {
 	 * <p>Generates the list of the enumerated types.</p>
 	 * @return the list of all possible values
 	 */
-	public static List<EnumVerbDefiniteness> asList() {
-		return Arrays.asList(INDEFINITE, DEFINITE);
+	public static List<EnumVerbTenses> asList() {
+		return Arrays.asList(IMPERFECT, PERFECT);
 	}
 	
 }
