+++ src/main/java/br/com/collei/lavi/morphology/enums/filter/GenericFilter.java
@@ -1,4 +1,4 @@
-package br.com.collei.lavi.morphology.filter;
+package br.com.collei.lavi.morphology.enums.filter;
 
 /**
  * <p>A generic filter, extended by NounDeclensionFilter and VerbConjugatedFilter.</p>
