+++ src/main/java/br/com/collei/lavi/morphology/enums/inflect/ConjugatedItem.java
@@ -1,13 +1,13 @@
-package br.com.collei.lavi.morphology.inflect;
+package br.com.collei.lavi.morphology.enums.inflect;
 
-import br.com.collei.lavi.morphology.EnumVerbDefiniteness;
-import br.com.collei.lavi.morphology.EnumVerbModes;
-import br.com.collei.lavi.morphology.EnumVerbPersons;
-import br.com.collei.lavi.morphology.EnumVerbTenses;
-import br.com.collei.lavi.morphology.EnumVerbVoices;
-import br.com.collei.lavi.morphology.core.Term;
-import br.com.collei.lavi.morphology.core.Verb;
-import br.com.collei.lavi.morphology.filter.VerbConjugationFilter;
+import br.com.collei.lavi.morphology.enums.EnumVerbDefiniteness;
+import br.com.collei.lavi.morphology.enums.EnumVerbModes;
+import br.com.collei.lavi.morphology.enums.EnumVerbPersons;
+import br.com.collei.lavi.morphology.enums.EnumVerbTenses;
+import br.com.collei.lavi.morphology.enums.EnumVerbVoices;
+import br.com.collei.lavi.morphology.enums.core.Term;
+import br.com.collei.lavi.morphology.enums.core.Verb;
+import br.com.collei.lavi.morphology.enums.filter.VerbConjugationFilter;
 
 /**
  * <p>A conjugated item in a ConjugationTable instance.</p>
