+++ src/test/java/br/com/collei/lavi/morphology/TermTest.java
@@ -6,8 +6,8 @@ import static org.junit.jupiter.api.Assertions.assertTrue;
 
 import org.junit.jupiter.api.Test;
 
-import br.com.collei.lavi.morphology.EnumHarmony;
-import br.com.collei.lavi.morphology.core.Term;
+import br.com.collei.lavi.morphology.enums.EnumHarmony;
+import br.com.collei.lavi.morphology.enums.core.Term;
 
 public class TermTest {
