+++ src/test/java/br/com/collei/lavi/morphology/inflect/ConjugationTableTest.java
@@ -8,16 +8,16 @@ import java.util.List;
 
 import org.junit.jupiter.api.Test;
 
-import br.com.collei.lavi.morphology.EnumVerbDefiniteness;
-import br.com.collei.lavi.morphology.EnumVerbModes;
-import br.com.collei.lavi.morphology.EnumVerbPersons;
-import br.com.collei.lavi.morphology.EnumVerbTenses;
-import br.com.collei.lavi.morphology.EnumVerbVoices;
-import br.com.collei.lavi.morphology.core.Term;
-import br.com.collei.lavi.morphology.core.Verb;
-import br.com.collei.lavi.morphology.filter.VerbConjugationFilter;
-import br.com.collei.lavi.morphology.inflect.ConjugatedItem;
-import br.com.collei.lavi.morphology.inflect.ConjugationTable;
+import br.com.collei.lavi.morphology.enums.EnumVerbDefiniteness;
+import br.com.collei.lavi.morphology.enums.EnumVerbModes;
+import br.com.collei.lavi.morphology.enums.EnumVerbPersons;
+import br.com.collei.lavi.morphology.enums.EnumVerbTenses;
+import br.com.collei.lavi.morphology.enums.EnumVerbVoices;
+import br.com.collei.lavi.morphology.enums.core.Term;
+import br.com.collei.lavi.morphology.enums.core.Verb;
+import br.com.collei.lavi.morphology.enums.filter.VerbConjugationFilter;
+import br.com.collei.lavi.morphology.enums.inflect.ConjugatedItem;
+import br.com.collei.lavi.morphology.enums.inflect.ConjugationTable;
 
 public class ConjugationTableTest {
