+++ src/test/java/br/com/collei/lavi/morphology/inflect/DeclensionTableTest.java
@@ -11,14 +11,14 @@ import java.util.List;
 
 import org.junit.jupiter.api.Test;
 
-import br.com.collei.lavi.morphology.EnumNounCases;
-import br.com.collei.lavi.morphology.EnumNounPersons;
-import br.com.collei.lavi.morphology.EnumNounPluralities;
-import br.com.collei.lavi.morphology.core.Noun;
-import br.com.collei.lavi.morphology.core.Term;
-import br.com.collei.lavi.morphology.filter.NounDeclensionFilter;
-import br.com.collei.lavi.morphology.inflect.DeclensedItem;
-import br.com.collei.lavi.morphology.inflect.DeclensionTable;
+import br.com.collei.lavi.morphology.enums.EnumNounCases;
+import br.com.collei.lavi.morphology.enums.EnumNounPersons;
+import br.com.collei.lavi.morphology.enums.EnumNounPluralities;
+import br.com.collei.lavi.morphology.enums.core.Noun;
+import br.com.collei.lavi.morphology.enums.core.Term;
+import br.com.collei.lavi.morphology.enums.filter.NounDeclensionFilter;
+import br.com.collei.lavi.morphology.enums.inflect.DeclensedItem;
+import br.com.collei.lavi.morphology.enums.inflect.DeclensionTable;
 
 public class DeclensionTableTest {
