+++ src/test/java/br/com/collei/lavi/morphology/test/helpers/VerbTestItem.java
@@ -1,12 +1,12 @@
 package br.com.collei.lavi.morphology.test.helpers;
 
-import br.com.collei.lavi.morphology.EnumVerbDefiniteness;
-import br.com.collei.lavi.morphology.EnumVerbModes;
-import br.com.collei.lavi.morphology.EnumVerbPersons;
-import br.com.collei.lavi.morphology.EnumVerbTenses;
-import br.com.collei.lavi.morphology.EnumVerbVoices;
-import br.com.collei.lavi.morphology.core.Term;
-import br.com.collei.lavi.morphology.core.Verb;
+import br.com.collei.lavi.morphology.enums.EnumVerbDefiniteness;
+import br.com.collei.lavi.morphology.enums.EnumVerbModes;
+import br.com.collei.lavi.morphology.enums.EnumVerbPersons;
+import br.com.collei.lavi.morphology.enums.EnumVerbTenses;
+import br.com.collei.lavi.morphology.enums.EnumVerbVoices;
+import br.com.collei.lavi.morphology.enums.core.Term;
+import br.com.collei.lavi.morphology.enums.core.Verb;
 
 public class VerbTestItem {
